Continue qt framework -> qt creator integration
Code between //@!ANDROID-4 will be commented for android-4 platform

+++ src/plugins/qt4projectmanager/qt-android/androidpackagecreationstep.cpp
@@ -139,15 +139,20 @@ bool AndroidPackageCreationStep::createPackage(QProcess *buildProc)
         {
             case ApplicationTemplate:
                 fileName=node->targetInformation().target;
-                androidFileName=QLatin1String("lib")+node->targetInformation().target+QLatin1String(".so");
-                QFile::remove(node->targetInformation().buildDir+QLatin1Char('/')+androidFileName);
-                if (!QFile::copy(node->targetInformation().buildDir+QLatin1Char('/')+fileName,
-                            node->targetInformation().buildDir+QLatin1Char('/')+androidFileName))
+                if (node->targetInformation().target.endsWith(QLatin1String(".so")))
+                    androidFileName=node->targetInformation().target;
+                else
                 {
-                    raiseError(tr("Cant copy '%1' from '%2' to '%3'").arg(fileName)
-                               .arg(node->targetInformation().buildDir)
-                               .arg(node->targetInformation().buildDir));
-                    return false;
+                    androidFileName=QLatin1String("lib")+node->targetInformation().target+QLatin1String(".so");
+                    QFile::remove(node->targetInformation().buildDir+QLatin1Char('/')+androidFileName);
+                    if (!QFile::copy(node->targetInformation().buildDir+QLatin1Char('/')+fileName,
+                                node->targetInformation().buildDir+QLatin1Char('/')+androidFileName))
+                    {
+                        raiseError(tr("Cant copy '%1' from '%2' to '%3'").arg(fileName)
+                                   .arg(node->targetInformation().buildDir)
+                                   .arg(node->targetInformation().buildDir));
+                        return false;
+                    }
                 }
                 break;
             case LibraryTemplate:
@@ -191,6 +196,9 @@ bool AndroidPackageCreationStep::createPackage(QProcess *buildProc)
     emit addOutput(tr("Creating package file ..."), MessageOutput);
     if (!AndroidTemplatesManager::instance()->createAndroidTemplatesIfNecessary(bc->qt4Target()->qt4Project()))
         return false;
+
+    AndroidTemplatesManager::instance()->updateProject(bc->qt4Target()->qt4Project(),
+                                                       AndroidTemplatesManager::instance()->targetSDK(bc->qt4Target()->qt4Project()));
 
     buildProc->setWorkingDirectory(androidDir);
